requirements.txt
flask
mongoengine
pymongo
python-dotenv
werkzeug
pyjwt
flask-cors
gunicorn
requests
bcrypt  
celery[redis]
redis
pytz
python-magic
